@@ -1,4 +1,4 @@
-package com.wix.RNSwipeView.events;
+package com.aveq.RNSwipeView.events;
 
 import com.facebook.react.uimanager.events.Event;
 import com.facebook.react.uimanager.events.RCTEventEmitter;
@@ -6,15 +6,15 @@ import com.facebook.react.uimanager.events.RCTEventEmitter;
 /**
  * Created by yedidyak on 07/09/2016.
  */
-public class BouncedBackEvent extends Event<BouncedBackEvent> {
+public class SwipedOutEvent extends Event<SwipedOutEvent> {
 
-    public BouncedBackEvent(int viewTag) {
+    public SwipedOutEvent(int viewTag) {
         super(viewTag);
     }
 
     @Override
     public String getEventName() {
-        return "onBouncedBack";
+        return "onSwipedOut";
     }
 
     @Override
